+++ main.go
@@ -25,7 +25,7 @@ import (
 
 // version is the current version of audit-cli.
 // Update this when releasing new versions following semantic versioning.
-const version = "0.1.0"
+const version = "0.2.0"
 
 func main() {
 	var rootCmd = &cobra.Command{